lunch_voter/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
flake8
pytest
pytest-django
